- Project restructured - #22 Added action for closing diff and search history tabs
